# Improve the performances of the flattening subcrate